requirements.txt
# A2A SDK
a2a-sdk>=0.2.0

# Model serving (self-hosted mode)
faster-whisper>=1.0.0
transformers>=4.40.0
vllm>=0.4.0
torch>=2.2.0

# API clients (API mode)
openai>=1.30.0
together>=1.0.0
google-generativeai>=0.5.0

# Instrumentation
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0

# Evaluation
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.27.0
uvicorn>=0.29.0
fastapi>=0.111.0
